+++ src/main/java/tk/themcbros/uselessmod/lists/ModFeatures.java
@@ -5,6 +5,7 @@ import java.util.List;
 import com.google.common.collect.Lists;
 
 import net.minecraft.util.ResourceLocation;
+import net.minecraft.world.gen.feature.BigTreeFeature;
 import net.minecraft.world.gen.feature.Feature;
 import net.minecraft.world.gen.feature.IFeatureConfig;
 import net.minecraft.world.gen.feature.NoFeatureConfig;
@@ -12,6 +13,7 @@ import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import tk.themcbros.uselessmod.UselessMod;
+import tk.themcbros.uselessmod.world.feature.UselessBigTreeFeature;
 import tk.themcbros.uselessmod.world.feature.UselessCropsFeature;
 import tk.themcbros.uselessmod.world.feature.UselessFlowersFeature;
 import tk.themcbros.uselessmod.world.feature.UselessTreeFeature;
@@ -24,6 +26,7 @@ public class ModFeatures {
 	public static final Feature<NoFeatureConfig> USELESS_FLOWERS = register("useless_flowers", new UselessFlowersFeature(NoFeatureConfig::deserialize));
 	public static final Feature<NoFeatureConfig> USELESS_CROPS = register("useless_crops", new UselessCropsFeature(NoFeatureConfig::deserialize));
 	public static final Feature<NoFeatureConfig> USELESS_TREES = register("useless_trees", new UselessTreeFeature(NoFeatureConfig::deserialize, false, false));
+	public static final Feature<NoFeatureConfig> FANCY_USELESS_TREE = register("fancy_useless_tree", new UselessBigTreeFeature(NoFeatureConfig::deserialize, false));
 	
 	private static final <T extends Feature<? extends IFeatureConfig>> T register(String regName, T feature) {
 		feature.setRegistryName(new ResourceLocation(UselessMod.MOD_ID, regName));