+++ src/main/java/tk/themcbros/uselessmod/lists/ModFluids.java
@@ -15,6 +15,7 @@ import net.minecraftforge.fluids.ForgeFlowingFluid;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.registries.ObjectHolder;
 import tk.themcbros.uselessmod.UselessMod;
+import tk.themcbros.uselessmod.fluids.UselessWaterFluid;
 
 @ObjectHolder(UselessMod.MOD_ID)
 @Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.MOD, modid = UselessMod.MOD_ID)
@@ -22,13 +23,13 @@ public class ModFluids {
 
 	private static final List<Fluid> FLUIDS = Lists.newArrayList();
 	
-	private static final ForgeFlowingFluid.Properties PROPERTIES = new ForgeFlowingFluid.Properties(() -> ModFluids.USELESS_WATER, () -> ModFluids.FLOWING_USELESS_WATER, 
-			FluidAttributes.builder( new ResourceLocation("block/water_still"), new ResourceLocation("block/water_flow")).overlay(new ResourceLocation("block/water_overlay"))
-                    .translationKey("block.uselessmod.useless_water")
-                    .color(0xFF468b44).density(1000))
-			.block(() -> ModBlocks.USELESS_WATER).bucket(() -> ModItems.USELESS_WATER_BUCKET).canMultiply().renderLayer(BlockRenderLayer.TRANSLUCENT);
-	public static final FlowingFluid USELESS_WATER = registerFluid("useless_water", new ForgeFlowingFluid.Source(PROPERTIES));
-	public static final FlowingFluid FLOWING_USELESS_WATER = registerFluid("flowing_useless_water", new ForgeFlowingFluid.Flowing(PROPERTIES));
+//	private static final ForgeFlowingFluid.Properties PROPERTIES = new ForgeFlowingFluid.Properties(() -> ModFluids.USELESS_WATER, () -> ModFluids.FLOWING_USELESS_WATER,
+//			FluidAttributes.builder( new ResourceLocation("block/water_still"), new ResourceLocation("block/water_flow")).overlay(new ResourceLocation("block/water_overlay"))
+//                    .translationKey("block.uselessmod.useless_water")
+//                    .color(0xFF468b44).density(1000))
+//			.block(() -> ModBlocks.USELESS_WATER).bucket(() -> ModItems.USELESS_WATER_BUCKET).canMultiply().renderLayer(BlockRenderLayer.TRANSLUCENT);
+	public static final FlowingFluid USELESS_WATER = registerFluid("useless_water", new UselessWaterFluid.Source());
+	public static final FlowingFluid FLOWING_USELESS_WATER = registerFluid("flowing_useless_water", new UselessWaterFluid.Flowing());
 	
 	private static <T extends Fluid> T registerFluid(String registryName, T fluid) {
 		fluid.setRegistryName(new ResourceLocation(UselessMod.MOD_ID, registryName));
@@ -38,7 +39,7 @@ public class ModFluids {
 
 	@SubscribeEvent
 	public static void onFluidRegister(final RegistryEvent.Register<Fluid> event) {
-		FLUIDS.forEach(fluid -> event.getRegistry().register(fluid));
+		FLUIDS.forEach(event.getRegistry()::register);
 		UselessMod.LOGGER.debug("Registered fluids");
 	}