+++ CHANGELOG.md
@@ -9,7 +9,9 @@
 - Added functionality to the Power Control Block ([?](https://github.com/TheMCBrothers/UselessMod/wiki/Blocks#Power-Control-Block))
 - Added Useless Soup recipe
 - Added machine tier upgrades
+- Added Big Useless trees
 - Useless Blocks and Super Useless Blocks can be used as beacon base
+- Generate useless water lakes in useless biome
 - Fixed Purple Lamp recipe
 - Fixed Energy Cables
 - Organized registries